+++ components/Footer/LongFooter.tsx
@@ -1,42 +1,13 @@
 /** @jsx jsx */
-import { Flex, Grid, Box, Input, Button, Container, jsx } from 'theme-ui';
+import { Flex, Grid, Box, Container, jsx, Text} from 'theme-ui';
 import { Icon } from '@makerdao/dai-ui-icons';
 
-import { subscribeToNewsletter } from 'lib/services/emailSubscription';
-import { isValidEmail } from 'lib/validators/email';
-import React, { useState } from 'react';
+import React from 'react';
 import { translate } from '@makerdao/i18n-helper';
 import Link from 'next/link';
 
-export default function LongFooter({ locale = 'en' }: { locale: string }): React.ReactElement {
-  const [email, setEmail] = useState('');
-  const [subscribeState, setSubscribeState] = useState('');
-  const [message, setMessage] = useState('');
-
-  const updateEmail = evt => {
-    setEmail(evt.target.value);
-    setMessage('');
-    setSubscribeState('');
-  };
-
-  const subscribeEmail = async () => {
-    if (!isValidEmail(email)) {
-      setMessage('Please enter a valid email address.');
-      setSubscribeState('failure');
-      return;
-    }
-
-    try {
-      await subscribeToNewsletter(email);
-
-      setSubscribeState('success');
-      setMessage("Thank you. You'll hear from us soon.");
-    } catch (e) {
-      setSubscribeState('failure');
-      setMessage('An error occurred. Please try again later.');
-    }
-  };
-
+export default function LongFooter({ locale = 'en' }: { locale?: string }): React.ReactElement {
+ 
   const t = text => translate(text, locale);
 
   const links = [
@@ -180,56 +151,10 @@ export default function LongFooter({ locale = 'en' }: { locale: string }): React
           </div>
         ))}
 
-        <Box gridColumn={['1 / 3', '1 / 3', 'unset']}>
-          <Flex sx={{ width: '284px', flexDirection: 'row' }} mt="1.2rem">
-            <Input
-              py="xs"
-              placeholder="Sign up for our newsletter"
-              sx={{
-                fontSize: 15,
-                borderRadius: '5px 0px 0px 5px',
-                '::placeholder': {
-                  color: 'formGrey',
-                  opacity: 1,
-                  fontWeight: '200'
-                },
-                ':focus': { borderColor: 'secondary' },
-                maxWidth: '320px'
-              }}
-              name="email"
-              type="email"
-              value={email}
-              onChange={evt => updateEmail(evt)}
-              errorMessage={(subscribeState === 'failure' || undefined) && message}
-              successMessage={(subscribeState === 'success' || undefined) && message}
-            />
-            <Button
-              variant="outline"
-              type="submit"
-              onClick={subscribeEmail}
-              sx={{
-                padding: '0px',
-                paddingTop: '4px',
-                width: '44px',
-                border: '1px solid',
-                borderColor: 'secondary',
-                borderLeft: 0,
-                borderRadius: '0px 5px 5px 0px',
-                ':hover': { borderColor: 'secondary' },
-                ':hover svg': { transform: 'translate(0.25rem)' },
-                svg: {
-                  margin: '0 auto',
-                  transition: 'transform 0.125s'
-                }
-              }}
-            >
-              <Icon name="subscribe_arrow" />
-            </Button>
-          </Flex>
+        <Box >
+          
           <Flex
-            gap="20px"
             mt="1.8rem"
-            columns="repeat(7, 20px)"
             sx={{
               flexDirection: 'row',
               justifyContent: 'left',
