+++ pages/polling/review.tsx
@@ -22,7 +22,6 @@ import useBallotStore from 'stores/ballot';
 import useAccountsStore from 'stores/accounts';
 import MobileVoteSheet from 'components/polling/MobileVoteSheet';
 import mixpanel from 'mixpanel-browser';
-import Skeleton from 'react-loading-skeleton';
 import PageLoadingPlaceholder from 'components/PageLoadingPlaceholder';
 
 const PollingReview = ({ polls }: { polls: Poll[] }) => {
